requirements.txt
numpy==1.26.4
transformers==4.41.2
huggingface-hub==0.23.4
tokenizers==0.19.1
torch==2.2.2
sentencepiece
pydantic==2.7.4
python-dotenv==1.0.1
rich==13.7.1
